requirements.txt
fastapi
uvicorn[standard]
openai>=1.0.0
faiss-cpu
numpy
httpx
python-dotenv